requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn[standard]==0.23.2
pybase62
pymongo==4.6.1